backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary>=2.9.10
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.9
google-generativeai==0.8.3
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
